requirements.txt
pywebview == 5.4
Flask == 3.1.0